move space commands under space from upbound
