+++ cmd/up/space/uninstall.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package upbound
+package space
 
 import (
 	"github.com/upbound/up/internal/install"
